+++ lib/block/minecraft/piston.ts
@@ -2,7 +2,10 @@ import { map } from "../../iterable.js";
 import { Block } from "../../block.js";
 import { Wrapper } from "../../wrapper.js";
 import { Faced } from "../faced.js";
+import { BlockPistonState } from "@minecraft/server";
 import * as MC from "@minecraft/server";
+
+export { BlockPistonState };
 
 export class Piston extends Faced(Block) {
     readonly #piston: MC.BlockPistonComponent;
@@ -14,24 +17,12 @@ export class Piston extends Faced(Block) {
             this.getComponentOrThrow(MC.BlockPistonComponent.componentId);
     }
 
-    public get isExpanded(): boolean {
-        return this.#piston.isExpanded;
-    }
-
-    public get isExpanding(): boolean {
-        return this.#piston.isExpanding;
-    }
-
     public get isMoving(): boolean {
         return this.#piston.isMoving;
     }
 
-    public get isRetracted(): boolean {
-        return this.#piston.isRetracted;
-    }
-
-    public get isRetracting(): boolean {
-        return this.#piston.isRetracting;
+    public get state(): BlockPistonState {
+        return this.#piston.state;
     }
 
     public get attachedBlocks(): IterableIterator<Block> {
@@ -50,20 +41,8 @@ export class PistonActivateAfterEvent extends Wrapper<MC.PistonActivateAfterEven
         super(rawEv);
         // Every field of MC.PistonActivateAfterEvent becomes undefined
         // after the event handler returns. We must copy them now.
+        // THINKME: That sounds like a game bug. Is that still the case?
         this.isExpanding = rawEv.isExpanding;
         this.piston      = new Piston(rawEv.piston.block);
-    }
-}
-
-export class PistonActivateBeforeEvent extends PistonActivateAfterEvent {
-    /// Package private
-    public constructor(rawEv: MC.PistonActivateBeforeEvent) {
-        super(rawEv);
-    }
-
-    public cancel(): void {
-        // NOTE: We cannot simply override "get raw()" due to
-        // https://github.com/microsoft/TypeScript/issues/41347
-        (super.raw as MC.PistonActivateBeforeEvent).cancel = true;
     }
 }